Break out top-level REST routers into separate files
