+++ vibin/server/routers/__init__.py
@@ -1,0 +1,12 @@
+from .albums import albums_router
+from .artists import artists_router
+from .browse import browse_router
+from .favorites import favorites_router
+from .playlist import playlist_router
+from .playlists import playlists_router
+from .presets import presets_router
+from .system import system_router
+from .tracks import tracks_router
+from .transport import transport_router
+from .vibin import vibin_router
+
