+++ vibin/server/routers/tracks.py
@@ -1,0 +1,192 @@
+import json
+import math
+from typing import List, Optional, Union
+
+from fastapi import APIRouter, Header, HTTPException, Response
+
+from vibin import Vibin, VibinMissingDependencyError, VibinNotFoundError
+from vibin.logger import logger
+from vibin.models import LyricsQuery, Track
+
+
+def tracks_router(vibin: Vibin, requires_media, transform_media_server_urls_if_proxying):
+    router = APIRouter()
+
+    @router.get("/tracks", summary="", description="", tags=["Tracks"])
+    @transform_media_server_urls_if_proxying
+    @requires_media
+    def tracks() -> List[Track]:
+        try:
+            return vibin.media.tracks
+        except VibinNotFoundError as e:
+            raise HTTPException(status_code=404, detail=str(e))
+
+    @router.get("/tracks/lyrics", summary="", description="", tags=["Tracks"])
+    def track_lyrics(artist: str, title: str, update_cache: Optional[bool] = False):
+        lyrics = vibin.lyrics_for_track(
+            artist=artist, title=title, update_cache=update_cache
+        )
+
+        if lyrics is None:
+            raise HTTPException(status_code=404, detail="Lyrics not found")
+
+        return lyrics
+
+    @router.post(
+        "/tracks/lyrics/validate", summary="", description="", tags=["Tracks"]
+    )
+    def track_lyrics_validate(artist: str, title: str, is_valid: bool):
+        lyrics = track_lyrics(artist=artist, title=title)
+        vibin.lyrics_valid(lyrics_id=lyrics["id"], is_valid=is_valid)
+
+        return track_lyrics(artist=artist, title=title)
+
+    @router.post(
+        "/tracks/lyrics/search", summary="", description="", tags=["Tracks"]
+    )
+    def track_lyrics_search(lyrics_query: LyricsQuery):
+        results = vibin.lyrics_search(lyrics_query.query)
+
+        return {
+            "query": lyrics_query.query,
+            "matches": results,
+        }
+
+    @router.get("/tracks/links", summary="", description="", tags=["Tracks"])
+    def track_links(
+        artist: Optional[str],
+        album: Optional[str],
+        title: Optional[str],
+        all_types: bool = False,
+    ):
+        return vibin.media_links(
+            artist=artist, album=album, title=title, include_all=all_types
+        )
+
+    @router.get("/tracks/{track_id}", summary="", description="", tags=["Tracks"])
+    @transform_media_server_urls_if_proxying
+    @requires_media
+    def track_by_id(track_id: str) -> Track:
+        try:
+            return vibin.media.track(track_id)
+        except VibinNotFoundError as e:
+            raise HTTPException(status_code=404, detail=str(e))
+
+    @router.get(
+        "/tracks/{track_id}/lyrics", summary="", description="", tags=["Tracks"]
+    )
+    def track_lyrics_by_track_id(track_id: str, update_cache: Optional[bool] = False):
+        lyrics = vibin.lyrics_for_track(track_id=track_id, update_cache=update_cache)
+
+        if lyrics is None:
+            raise HTTPException(status_code=404, detail="Lyrics not found")
+
+        return lyrics
+
+    @router.post(
+        "/tracks/{track_id}/lyrics/validate",
+        summary="",
+        description="",
+        tags=["Tracks"],
+    )
+    def track_lyrics_by_track_id_validate(track_id: str, is_valid: bool):
+        lyrics = track_lyrics_by_track_id(track_id)
+        vibin.lyrics_valid(lyrics_id=lyrics.lyrics_id, is_valid=is_valid)
+
+        return track_lyrics_by_track_id(track_id)
+
+    @router.get(
+        "/tracks/{track_id}/waveform.png", summary="", description="", tags=["Tracks"]
+    )
+    def track_waveform_png(
+        track_id: str,
+        width: int = 800,
+        height: int = 250,
+    ):
+        try:
+            waveform = vibin.waveform_for_track(
+                track_id, data_format="png", width=width, height=height
+            )
+
+            return Response(content=waveform, media_type="image/png")
+        except VibinMissingDependencyError as e:
+            # TODO: Where possible, have errors reference docs for possible
+            #   actions the caller can take to resolve the issue.
+            logger.warning(f"Cannot generate waveform due to missing dependency: {e}")
+
+            raise HTTPException(
+                status_code=404,
+                detail=f"Cannot generate waveform due to missing dependency: {e}",
+            )
+
+    @router.get(
+        "/tracks/{track_id}/waveform", summary="", description="", tags=["Tracks"]
+    )
+    def track_waveform(
+        track_id: str,
+        width: int = 800,
+        height: int = 250,
+        accept: Union[str, None] = Header(default="application/json"),
+    ):
+        # TODO: This waveform_format / media_type / "accept" header stuff
+        #   feels too convoluted.
+        waveform_format = "json"
+        media_type = "application/json"
+
+        if accept == "application/octet-stream":
+            waveform_format = "dat"
+            media_type = "application/octet-stream"
+        elif accept == "image/png":
+            waveform_format = "png"
+            media_type = "image/png"
+
+        try:
+            if waveform_format == "png":
+                waveform = vibin.waveform_for_track(
+                    track_id, data_format=waveform_format, width=width, height=height
+                )
+            else:
+                waveform = vibin.waveform_for_track(
+                    track_id, data_format=waveform_format
+                )
+
+            return Response(
+                content=json.dumps(waveform) if waveform_format == "json" else waveform,
+                media_type=media_type,
+            )
+        except VibinMissingDependencyError as e:
+            # TODO: Where possible, have errors reference docs for possible
+            #   actions the caller can take to resolve the issue.
+            raise HTTPException(
+                status_code=404,
+                detail=f"Cannot generate waveform due to missing dependency: {e}",
+            )
+
+    @router.get(
+        "/tracks/{track_id}/rms", summary="", description="", tags=["Tracks"]
+    )
+    def track_rms(track_id: str):
+        waveform = vibin.waveform_for_track(track_id, data_format="json")
+
+        samples = waveform["data"]
+        squared_samples = [sample**2 for sample in samples]
+        squared_sum = sum(squared_samples)
+
+        mean = squared_sum / len(samples)
+        rms = math.sqrt(mean)
+        peak = max((abs(sample) for sample in samples))
+        rms_to_peak_ratio = rms / peak
+
+        return {
+            "rms": rms,
+            "peak": peak,
+            "rms_to_peak": rms_to_peak_ratio,
+        }
+
+    @router.get(
+        "/tracks/{track_id}/links", summary="", description="", tags=["Tracks"]
+    )
+    def track_links_by_track_id(track_id: str, all_types: bool = False):
+        return vibin.media_links(media_id=track_id, include_all=all_types)
+
+    return router
